pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]
